Remove blanket recomomendation to verify computed CSS
CSS is a tool for modifying the presentation of a website.
Asserting whether a CSS class is present doesn't give strong
confidence that a webpage is being presented correctly.
Worse, it makes refactoring class names more painful.

To get confidence that the presentation of a website is correct is
best done either by a human visually inspecting the page, or an
automated visual regression tool.  Neither of which will hinder
refactoring.

I can imagine some scenarios where asserting against the style
of something is important, but these would be outlier cases.