Create CODEOWNERS
This change would automatically ask for a review from the dfe-lead-devs team when a PR is raised. I think this will help us control the flow of information into the repo and make sure our standards, principles, and guidance are technically sound. It would also allow us to give write access to everyone in dfe-digital, thereby making it much easier to raise a PR and start contributing to the repo.

It should be combined with the GitHub interface flag to ensure that PRs can only be merged when they've been approved by someone in this team